feat: finish basic implement. and reduce coupling
Players now take turns to attack the opponent's
board. The UI shows whose turn is at the moment.
If a player wins, the UI shows the winner and
it is impossible to make further turns.

The project was too tightly coupled, so I
refactored it to abide by the Law of Demeter.

Now, I need to make it possible to place ships
on the board in the UI.