fix: make old sqs undraggable, new sqs draggable
Because of some mistakes in code, when you dragged
a ship and dropped it somewhere, the old squares
were still draggable, and the new ones... they
were, but I guess it's just because the browser
figured out that they had to be draggable. And it
was far from ideal. Now, everything works as
intended.